+++ dom/src/main/java/react4j/dom/proptypes/html/RefConsumer.java
@@ -1,5 +1,6 @@
 package react4j.dom.proptypes.html;
 
+import javax.annotation.Nullable;
 import jsinterop.annotations.JsFunction;
 
 /**
@@ -11,8 +12,10 @@ public interface RefConsumer
 {
   /**
    * Passes the reference to the component instance or element.
+   * The reference is nonnull when the element has been attached to the DOM and
+   * null when the reference has been detached from the DOM.
    *
    * @param reference the reference.
    */
-  void accept( Object reference );
+  void accept( @Nullable Object reference );
 }